+++ packages/data-prefetch/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @module-federation/data-prefetch
 
+## 0.24.0
+
+### Patch Changes
+
+- @module-federation/runtime@0.24.0
+- @module-federation/sdk@0.24.0
+
 ## 0.23.0
 
 ### Patch Changes