+++ packages/node/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @module-federation/node
 
+## 2.7.29
+
+### Patch Changes
+
+- Updated dependencies [c3c068e]
+  - @module-federation/enhanced@0.24.0
+  - @module-federation/runtime@0.24.0
+  - @module-federation/sdk@0.24.0
+
 ## 2.7.28
 
 ### Patch Changes