+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @module-federation/runtime
 
+## 0.24.0
+
+### Patch Changes
+
+- @module-federation/sdk@0.24.0
+- @module-federation/error-codes@0.24.0
+- @module-federation/runtime-core@0.24.0
+
 ## 0.23.0
 
 ### Patch Changes